# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Twitter Fullstack Exam API",
        "description": "REST API used by the frontend. Sanctum bearer tokens required for all protected endpoints.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/comments": {
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Create a new comment",
                "description": "Creates a new comment on a post and returns the created comment with user information.",
                "operationId": "300619a0a4003920aafe960e45688102",
                "requestBody": {
                    "description": "Comment content and post reference",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "post_pk",
                                    "comment_message"
                                ],
                                "properties": {
                                    "post_pk": {
                                        "description": "Primary key of the post to comment on",
                                        "type": "string",
                                        "example": "post123"
                                    },
                                    "comment_message": {
                                        "type": "string",
                                        "maxLength": 280,
                                        "example": "This is a comment!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comment created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The selected post_pk is invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error (e.g., missing fields or invalid post reference)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error while creating the comment",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/comments/{comment_pk}": {
            "put": {
                "tags": [
                    "Comments"
                ],
                "summary": "Update a comment",
                "description": "Updates the message of an existing comment and returns the updated comment with user information.",
                "operationId": "01afd98d1693be7c0cc35a7255caa8a9",
                "parameters": [
                    {
                        "name": "comment_pk",
                        "in": "path",
                        "description": "Primary key of the comment to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated comment message",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "comment_message"
                                ],
                                "properties": {
                                    "comment_message": {
                                        "type": "string",
                                        "maxLength": 280,
                                        "example": "This is the updated comment."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Comment updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Comment not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Comment not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error (e.g., missing or invalid message)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error while updating the comment",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Delete a comment",
                "description": "Permanently deletes a comment identified by its primary key.",
                "operationId": "bfe49a5eca4a038bd125ff6522b08091",
                "parameters": [
                    {
                        "name": "comment_pk",
                        "in": "path",
                        "description": "Primary key of the comment to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comment deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Comment deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Comment not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Comment not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error while deleting the comment",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "properties": {
                    "user_pk": {
                        "type": "string",
                        "example": "user123"
                    },
                    "user_username": {
                        "type": "string",
                        "example": "johndoe"
                    },
                    "user_email": {
                        "type": "string",
                        "example": "john@example.com"
                    },
                    "user_full_name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "user_profile_picture": {
                        "type": "string",
                        "example": "profile_pictures/user123.jpg",
                        "nullable": true
                    },
                    "user_cover_picture": {
                        "type": "string",
                        "example": "cover_pictures/user123.jpg",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Comment": {
                "properties": {
                    "comment_pk": {
                        "type": "string",
                        "example": "comment123"
                    },
                    "comment_message": {
                        "type": "string",
                        "example": "This is a comment!"
                    },
                    "comment_post_fk": {
                        "type": "string",
                        "example": "post123"
                    },
                    "comment_user_fk": {
                        "type": "string",
                        "example": "user123"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}